Input: synaptics-rmi4 - Introduce a new feature for Forcepad.
Forcepad devices will replace metal buttons with vibrators.
Therefore, the RMI4 driver will use the F21 function to
simulate whether the metal button is pressed.
When the fingertip's weight exceeds the default value,
the F21 function will indicate that the click button
is pressed.
We will use the F21 function to replace the existing
F30 and F3A functions on the Forcepad device.

Signed-off-by: Marge Yang <marge.yang@tw.synaptics.com>
Signed-off-by: Vincent Huang <vincent.huang@tw.synaptics.com>